requirements.txt
gradio
transformers
torchvision
PySide6-Fluent-Widgets[full]
